+++ src/xso/__init__.py
@@ -30,4 +30,4 @@ from .component import component
 
 from .variables import variable, parameter, forcing, flux, index
 
-from .xsimlabwrappers import create, setup
+from .xsimlabwrappers import create, setup, update_setup